pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argus"
version = "0.1.0"
description = "Automated multi-strategy day trading ecosystem"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5,<3",
    "PyYAML>=6.0,<7",
    "aiosqlite>=0.19,<1",
    "python-ulid>=2.2,<3",
    "pandas>=2.0,<3",
    "pyarrow>=14.0,<18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=4.1,<6",
    "ruff>=0.2,<1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.setuptools.packages.find]
include = ["argus*"]
